+++ packages/amplify-e2e-core/src/init/amplifyPush.ts
@@ -1,8 +1,10 @@
 import { nspawn as spawn, getCLIPath } from '../../src';
 
+const pushTimeoutMS = 1000 * 60 * 10; // 10 minutes;
+
 export function amplifyPush(cwd: string, testingWithLatestCodebase: boolean = false) {
   return new Promise((resolve, reject) => {
-    spawn(getCLIPath(testingWithLatestCodebase), ['push'], { cwd, stripColors: true })
+    spawn(getCLIPath(testingWithLatestCodebase), ['push'], { cwd, stripColors: true, noOutputTimeout: pushTimeoutMS })
       .wait('Are you sure you want to continue?')
       .sendLine('y')
       .wait('Do you want to generate code for your newly created GraphQL API')
@@ -20,7 +22,7 @@ export function amplifyPush(cwd: string, testingWithLatestCodebase: boolean = fa
 
 export function amplifyPushWithoutCodegen(cwd: string) {
   return new Promise((resolve, reject) => {
-    spawn(getCLIPath(), ['push'], { cwd, stripColors: true })
+    spawn(getCLIPath(), ['push'], { cwd, stripColors: true, noOutputTimeout: pushTimeoutMS })
       .wait('Are you sure you want to continue?')
       .sendCarriageReturn()
       .run((err: Error) => {
@@ -35,7 +37,7 @@ export function amplifyPushWithoutCodegen(cwd: string) {
 
 export function amplifyPushUpdate(cwd: string, waitForText?: RegExp, testingWithLatestCodebase: boolean = false) {
   return new Promise((resolve, reject) => {
-    spawn(getCLIPath(testingWithLatestCodebase), ['push'], { cwd, stripColors: true })
+    spawn(getCLIPath(testingWithLatestCodebase), ['push'], { cwd, stripColors: true, noOutputTimeout: pushTimeoutMS })
       .wait('Are you sure you want to continue?')
       .sendLine('y')
       .wait(waitForText || /.*/)
@@ -51,7 +53,7 @@ export function amplifyPushUpdate(cwd: string, waitForText?: RegExp, testingWith
 
 export function amplifyPushAuth(cwd: string, testingWithLatestCodebase: boolean = false) {
   return new Promise((resolve, reject) => {
-    spawn(getCLIPath(testingWithLatestCodebase), ['push'], { cwd, stripColors: true })
+    spawn(getCLIPath(testingWithLatestCodebase), ['push'], { cwd, stripColors: true, noOutputTimeout: pushTimeoutMS })
       .wait('Are you sure you want to continue?')
       .sendLine('y')
       .wait(/.*/)
