+++ packages/amplify-e2e-core/src/utils/sdk-calls.ts
@@ -10,19 +10,20 @@ import {
   CloudWatchLogs,
   CloudWatchEvents,
   Kinesis,
+  CloudFormation,
 } from 'aws-sdk';
 
-const getDDBTable = async (tableName: string, region: string) => {
+export const getDDBTable = async (tableName: string, region: string) => {
   const service = new DynamoDB({ region });
   return await service.describeTable({ TableName: tableName }).promise();
 };
 
-const checkIfBucketExists = async (bucketName: string, region: string) => {
+export const checkIfBucketExists = async (bucketName: string, region: string) => {
   const service = new S3({ region });
   return await service.headBucket({ Bucket: bucketName }).promise();
 };
 
-const getUserPool = async (userpoolId, region) => {
+export const getUserPool = async (userpoolId, region) => {
   config.update({ region });
   let res;
   try {
@@ -33,7 +34,7 @@ const getUserPool = async (userpoolId, region) => {
   return res;
 };
 
-const getLambdaFunction = async (functionName, region) => {
+export const getLambdaFunction = async (functionName, region) => {
   const lambda = new Lambda();
   let res;
   try {
@@ -44,7 +45,7 @@ const getLambdaFunction = async (functionName, region) => {
   return res;
 };
 
-const getUserPoolClients = async (userPoolId: string, clientIds: string[], region: string) => {
+export const getUserPoolClients = async (userPoolId: string, clientIds: string[], region: string) => {
   const provider = new CognitoIdentityServiceProvider({ region });
   const res = [];
   try {
@@ -63,47 +64,54 @@ const getUserPoolClients = async (userPoolId: string, clientIds: string[], regio
   return res;
 };
 
-const getBot = async (botName: string, region: string) => {
+export const getBot = async (botName: string, region: string) => {
   const service = new LexModelBuildingService({ region });
   return await service.getBot({ name: botName, versionOrAlias: '$LATEST' }).promise();
 };
 
-const getFunction = async (functionName: string, region: string) => {
+export const getFunction = async (functionName: string, region: string) => {
   const service = new Lambda({ region });
   return await service.getFunction({ FunctionName: functionName }).promise();
 };
 
-const invokeFunction = async (functionName: string, payload: string, region: string) => {
+export const invokeFunction = async (functionName: string, payload: string, region: string) => {
   const service = new Lambda({ region });
   return await service.invoke({ FunctionName: functionName, Payload: payload }).promise();
 };
 
-const getCollection = async (collectionId: string, region: string) => {
+export const getCollection = async (collectionId: string, region: string) => {
   const service = new Rekognition({ region });
   return await service.describeCollection({ CollectionId: collectionId }).promise();
 };
 
-const getTable = async (tableName: string, region: string) => {
+export const getTable = async (tableName: string, region: string) => {
   const service = new DynamoDB({ region });
   return await service.describeTable({ TableName: tableName }).promise();
 };
 
-const deleteTable = async (tableName: string, region: string) => {
+export const getEventSourceMappings = async (functionName: string, region: string) => {
+  const service = new Lambda({ region });
+  return (await service.listEventSourceMappings({ FunctionName: functionName }).promise()).EventSourceMappings;
+};
+
+export const deleteTable = async (tableName: string, region: string) => {
   const service = new DynamoDB({ region });
   return await service.deleteTable({ TableName: tableName }).promise();
 };
 
-const getAppSyncApi = async (appSyncApiId: string, region: string) => {
+export const getAppSyncApi = async (appSyncApiId: string, region: string) => {
   const service = new AppSync({ region });
   return await service.getGraphqlApi({ apiId: appSyncApiId }).promise();
 };
 
-const getCloudWatchLogs = async (region: string, logGroupName: string, logStreamName: string | undefined = undefined) => {
+export const getCloudWatchLogs = async (region: string, logGroupName: string, logStreamName: string | undefined = undefined) => {
   const cloudwatchlogs = new CloudWatchLogs({ region, retryDelayOptions: { base: 500 } });
 
   let targetStreamName = logStreamName;
   if (targetStreamName === undefined) {
-    const describeStreamsResp = await cloudwatchlogs.describeLogStreams({ logGroupName, descending: true }).promise();
+    const describeStreamsResp = await cloudwatchlogs
+      .describeLogStreams({ logGroupName, descending: true, orderBy: 'LastEventTime' })
+      .promise();
     if (describeStreamsResp.logStreams === undefined || describeStreamsResp.logStreams.length == 0) {
       return [];
     }
@@ -115,7 +123,12 @@ const getCloudWatchLogs = async (region: string, logGroupName: string, logStream
   return logsResp.events || [];
 };
 
-const putKinesisRecords = async (data: string, partitionKey: string, streamName: string, region: string) => {
+export const describeCloudFormationStack = async (stackName: string, region: string) => {
+  const service = new CloudFormation({ region });
+  return (await service.describeStacks({ StackName: stackName }).promise()).Stacks.find(stack => stack.StackName === stackName);
+};
+
+export const putKinesisRecords = async (data: string, partitionKey: string, streamName: string, region: string) => {
   const kinesis = new Kinesis({ region });
 
   return await kinesis
@@ -131,7 +144,7 @@ const putKinesisRecords = async (data: string, partitionKey: string, streamName:
     .promise();
 };
 
-const getCloudWatchEventRule = async (targetName: string, region: string) => {
+export const getCloudWatchEventRule = async (targetName: string, region: string) => {
   config.update({ region });
   const service = new CloudWatchEvents();
   var params = {
@@ -144,22 +157,4 @@ const getCloudWatchEventRule = async (targetName: string, region: string) => {
     console.log(e);
   }
   return ruleName;
-};
-
-export {
-  getDDBTable,
-  checkIfBucketExists,
-  getUserPool,
-  getUserPoolClients,
-  getBot,
-  getLambdaFunction,
-  getFunction,
-  invokeFunction,
-  getTable,
-  deleteTable,
-  getAppSyncApi,
-  getCollection,
-  getCloudWatchLogs,
-  putKinesisRecords,
-  getCloudWatchEventRule,
 };
